requirements.txt
flask
mysql-connector-python
flask-mail
flask-bcrypt
python-dateutil